+++ ohrms_holidays_approval/__manifest__.py
@@ -3,9 +3,9 @@
 {
     'name': 'Open HRMS Leave Multi-Level Approval',
     'version': '13.0.1.0.0',
-    'summary': """Multilevel approval for leaves""",
-    'description': 'leave approval, multiple leave approvers',
-    'category': 'Human Resources',
+    'summary': """Multilevel Approval for Leaves""",
+    'description': 'Multilevel Approval for Leaves, leave approval, multiple leave approvers, leave, approval',
+    'category': 'Generic Modules/Human Resources',
     'author': 'Cybrosys Techno Solutions',
     'company': 'Cybrosys Techno Solutions',
     'website': "https://www.openhrms.com",