+++ ohrms_holidays_approval/doc/RELEASE_NOTES.md
@@ -3,4 +3,4 @@
 #### 28.02.2020
 #### Version 13.0.1.0.0
 ##### ADD
-- Initial commit for Open HRMS multi level leave approval v12
+- Initial commit for Open HRMS multi level leave approval v13